ci fix
