+++ signal/micro/kernels/fft_auto_scale_common.cc
@@ -12,7 +12,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-#include "signal/micro/kernels/fft_auto_scale.h"
+#include "signal/micro/kernels/fft_auto_scale_kernel.h"
 #include "tensorflow/lite/kernels/kernel_util.h"
 #include "tensorflow/lite/micro/kernels/kernel_util.h"
